+++ demo/src/main/java/com/carlos/cutils/demo/MainActivity.kt
@@ -6,6 +6,7 @@ import androidx.fragment.app.Fragment
 import com.carlos.cutils.base.adapter.CBaseMyPagerAdapter
 import com.carlos.cutils.demo.fragment.ClipFragment
 import com.carlos.cutils.demo.fragment.RewardFragment
+import com.carlos.cutils.demo.fragment.ViewFragment
 import com.carlos.cutils.demo.fragment.WebviewFragment
 import com.carlos.cutils.listener.PermissionListener
 import com.carlos.cutils.util.LogUtils
@@ -17,8 +18,8 @@ class MainActivity : BaseActivity() {
 
     private val WECHAT_SERVICE_NAME = "com.carlos.grabredenvelope/.services.WechatService"
 
-    var fragments = mutableListOf<Fragment>( RewardFragment(), WebviewFragment(), ClipFragment())
-    var titles = mutableListOf("打赏", "网页", "剪贴")
+    var fragments = mutableListOf<Fragment>( RewardFragment(), WebviewFragment(), ClipFragment(), ViewFragment())
+    var titles = mutableListOf("打赏", "网页", "剪贴", "View")
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -27,6 +28,7 @@ class MainActivity : BaseActivity() {
         val adapter = CBaseMyPagerAdapter(supportFragmentManager, fragments, titles)
         viewpager.adapter = adapter
         sliding_tabs.setupWithViewPager(viewpager)
+        viewpager.currentItem = fragments.size-1
 
         getPermissions()