+++ demo/src/main/java/com/carlos/cutils/demo/fragment/ViewFragment.kt
@@ -1,0 +1,8 @@
+package com.carlos.cutils.demo.fragment
+
+import com.carlos.cutils.demo.R
+
+/**
+ * Created by Carlos on 2020/2/26.
+ */
+class ViewFragment : BaseFragment(R.layout.fragment_view)